+++ sentry-android-core/src/test/java/io/sentry/android/core/ManifestMetadataReaderTest.kt
@@ -381,4 +381,33 @@ class ManifestMetadataReaderTest {
         // Assert
         assertTrue(options.isEnableUncaughtExceptionHandler)
     }
+
+    @Test
+    fun `applyMetadata reads attachThreads to options`() {
+        // Arrange
+        val options = SentryAndroidOptions()
+        val bundle = Bundle()
+        val mockContext = ContextUtilsTest.mockMetaData(metaData = bundle)
+        bundle.putBoolean(ManifestMetadataReader.ATTACH_THREADS, true)
+
+        // Act
+        ManifestMetadataReader.applyMetadata(mockContext, options)
+
+        // Assert
+        assertTrue(options.isAttachThreads)
+    }
+
+    @Test
+    fun `applyMetadata reads attachThreads and keep default value if not found`() {
+        // Arrange
+        val options = SentryAndroidOptions()
+        val bundle = Bundle()
+        val mockContext = ContextUtilsTest.mockMetaData(metaData = bundle)
+
+        // Act
+        ManifestMetadataReader.applyMetadata(mockContext, options)
+
+        // Assert
+        assertFalse(options.isAttachThreads)
+    }
 }